RUMM-670 Fix the active span concept to let create several root spans on the same thread
A span is not automatically set as the activeSpan when it is created, there is a new method `setActive()` that must be called explicitly in the span for that functionality.
Any span still will use the current activeSpan as its parent on creation, except if any other explicit parent is assigned.