RUMM-1199 Fix "Undefined symbols for architecture arm64" in `carthage update --use-xcframeworks`
when installing `dd-sdk-ios` in other app. We need this as `arm64` is disabled in PLCrashReporter
simulator builds, so we cannot produce `DatadogCrashReporting.xcframework` as it is not able to
link `CrashReporter` to our `ios-arm64_x86_64-simulator` slice. This change makes
our SDK produce smaller slice, which is enough for running on simulator.

This was discovered as Carthage doesn’t yet support pre-build xcframeworks:
see: Carthage PR #3123