RUMM-1374 Disable `trailing_closure` linter rule as it doesn't fit our partial mocking convention
in tests, and it already requires 12 exclusions in the production code. After upgrading to Xcode 12.5
this rule started emitting 38 issues in our test module. Aligning the code to this rule would
result with unreadable partial mocking calls, so it's better to disable it entirely.

It started being visible only after upgrading to Xcode 12.5 - which ships new version of SourceKit.