RUMM-1934 Add option for including Xcode version in produced GH asset
this is required as due to broken Module Stability in XCFrameworks we need to ship
distinct XCFrameworks for each supported Xcode version. Module Stability is blocked
by the bug in Swift compiler: https://bugs.swift.org/browse/SR-14195. It will be fixed
in SDK V2 by avoiding names conflict that cause this issue.