Fix Xcode warning on linter build phase
The one:
Run script build phase '⚙️ Run linter' will be run during every build because it
does not specify any outputs. To address this warning, either add output dependencies
to the script phase, or configure it to run in every build by unchecking "Based on
dependency analysis" in the script phase.