RUMM-2950 Fix log attribute forwarding to RUM
When logs started going through the message bus, we accidentally started dropping the additional log attributes. This fixes it so we send the attributes properly. Additionally, we need to decode certain values specifically from the cross platform SDKs to ensure they're being set on the RUMErrorModel properly (sourceType and isCrash)