RUMM-3133 Don't override ViewScope attributes with UserActionScope attributes
### What and why?

Today, when a `UserAction`, `Resource` or `LongTask` event is sent with attributes attached, if the attributes have
conflicting keys with View event, the View attributes are overwritten by the `UserAction`/`Resource`/`LongTask`
attributes.

`UserActionScope` events must not update the `ViewScope` attributes, which has been the case for Android SDK.

### How?

Updated logic that overwrites the `ViewScope` attributes with `UserActionScope` attributes with unit tests to ensure
that the `ViewScope` attributes are not overwritten.

Fixed `Resource` events where parent attributes were reported when there is a conflict.