RUMM-3266 fix: use arch conditionals to assert arch specific signal
### What and why?

On x86 a `fatalError()` results on a SIGILL signal but on an ARM machine it results in a `SIGTRAP`.

Hence, the test case fail on the ARM machine. In order to update the CI to ARM, this needs to be fixed.

### How?

Use preprocessor macros to check the architecture and then use the appropriate signal.