RUMM-3222 Don't scale down refresh rate multiple times
### What and why?

During fixing the variable refresh rate calculation, the normalization was applied during the refresh rate determination phase, but the logic to normalize it again during the reporting was kept which resulted in double normalization.

Eg. First we normalize based on the logic of targetTimestamp and then divide by 2 for 120Hz display.

### How?

Remove second normalization phase during the reporting.