RUMM-3151 feat: reduce number of view updates by filtering events from payload
### What and why?

There are several view updates happens which are not needed. These view updates are reduced by the backend anyway, but
we could reduce the number of view updates by filtering them out from the payload.

### How?

While saving RUMViewEvent to disk, we save the metadata along with it. This allows us to access key information without
decoding the whole event. While preparing the request body, we filter out the events which are duplicate and will be
sent in the same batch.