RUMS-1778 Do not propagate Error and Long Task attributes to the View
A client raised an issue that Errors would get attributes from previous errors, which was making it hard for them to query on errors. It turns out this is because we were propagating attributes of errors up to the view.

The seems like this accidentally slipped in as a side affect as there were no tests checking for it, but I'm labeling it as a breaking change in the changelog.