feat: Add `build_id` to RUM and Logs
This adds a `build_id` to RUM and log events, mostly for cross platform frameworks that intend to use it as an identifier to assist for symbol file matching.

`build_id` is set on the Datadog Core context during initialization through additional properties and then added to Log and RUM events.

refs: RUM-2464