feat: Support adding binary images to non-crashes
This adds a cross platform attribute `_dd.error.include_binary_images` that will attach binary images to error reports by creating a backtrace and pulling the binary images from it.

It contains a TODO to replace backtrace generation with a different method of binary image attachment at a future time.