fix: Ensure need attributes are sent from logs on the message bus
Certain attributes (fingerprint and error source type) are removed as part of log processing because they are moved onto official properties of errors in logs.  However, when these errors are transmitted to RUM, the properties are not replaced, which they should be, so that RUM has all of the information about the error.